epocha/apps/users/views.py
"""Views for the users app."""
from rest_framework import generics, permissions

from .models import User
from .serializers import UserSerializer


class UserMeView(generics.RetrieveUpdateAPIView):
    """Authenticated user profile."""

    serializer_class = UserSerializer
    permission_classes = [permissions.IsAuthenticated]

    def get_object(self):
        return self.request.user
