requirements/base.txt
# Django
django>=5.1,<5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
django-environ>=0.11

# Channels (WebSocket)
channels>=4.1
channels-redis>=4.2
daphne>=4.1

# Celery
celery>=5.4
django-celery-beat>=2.6
redis>=5.0

# Database
psycopg[binary]>=3.2

# LLM
openai>=1.50

# Document parsing
pymupdf>=1.24
beautifulsoup4>=4.12
requests>=2.32
